Include Bukkit as well as CraftBukkit with higher depend to overwrite backwards compatibility layer in CraftBukkit..

+++ src/no/runsafe/framework/internal/wrapper/entity/BukkitLivingEntity.java
@@ -46,7 +46,7 @@ public abstract class BukkitLivingEntity extends RunsafeEntity
 		return ObjectWrapper.convert(this.entity.getEquipment());
 	}
 
-	public int getMaxHealth()
+	public double getMaxHealth()
 	{
 		return entity.getMaxHealth();
 	}
